refactor(planning): transition to new planning structure and algorithms
