+++ src/planning/CMakeLists.txt
@@ -45,18 +45,12 @@ set(IMPLEMENTATION_FILES
   src/adapter_planning/pacsim.cpp
   src/adapter_planning/vehicle.cpp
   src/utils/message_converter.cpp
-  src/planning/path_calculation2.cpp
+  src/planning/path_calculation.cpp
   src/planning/cone_coloring.cpp
-  src/planning/cone_coloring2.cpp
-  src/planning/smoothing2.cpp
+  src/planning/smoothing.cpp
   src/planning/outliers.cpp
-  src/planning/local_path_planner.cpp
-  src/planning/track.cpp
   src/planning/planning.cpp
-  src/planning/path_smoothing.cpp
   src/utils/cone.cpp
-  src/utils/cone2.cpp
-  src/utils/pathpoint.cpp
   src/utils/files.cpp
 )
