+++ src/planning/include/config/cone_coloring_config.hpp
@@ -1,5 +1,5 @@
-#ifndef SRC_PLANNING_INCLUDE_UTILS_CONE_COLORING_CONFIG_HPP_
-#define SRC_PLANNING_INCLUDE_UTILS_CONE_COLORING_CONFIG_HPP_
+#ifndef SRC_PLANNING_INCLUDE_CONFIG_CONE_COLORING_CONFIG_HPP_
+#define SRC_PLANNING_INCLUDE_CONFIG_CONE_COLORING_CONFIG_HPP_
 
 /**
  * @brief struct for the configuration of the cone coloring algorithm.
@@ -52,4 +52,4 @@ struct ConeColoringConfig {
         max_cost(max_cost) {}
 };
 
-#endif  // SRC_PLANNING_INCLUDE_UTILS_CONE_COLORING_CONFIG_HPP_
+#endif  // SRC_PLANNING_INCLUDE_CONFIG_CONE_COLORING_CONFIG_HPP_
