+++ src/planning/include/config/outliers_config.hpp
@@ -1,5 +1,5 @@
-#ifndef SRC_PLANNING_INCLUDE_UTILS_OUTLIERS_CONFIG_HPP_
-#define SRC_PLANNING_INCLUDE_UTILS_OUTLIERS_CONFIG_HPP_
+#ifndef SRC_PLANNING_INCLUDE_CONFIG_OUTLIERS_CONFIG_HPP_
+#define SRC_PLANNING_INCLUDE_CONFIG_OUTLIERS_CONFIG_HPP_
 
 /**
  * @brief struct for the configuration of the outliers removal algorithm.
@@ -26,4 +26,4 @@ struct OutliersConfig {
       : precision(precision), order(order), coeffs_ratio(coeffs_ratio) {}
 };
 
-#endif  // SRC_PLANNING_INCLUDE_UTILS_OUTLIERS_CONFIG_HPP_
+#endif  // SRC_PLANNING_INCLUDE_CONFIG_OUTLIERS_CONFIG_HPP_
