+++ src/planning/include/config/smoothing_config.hpp
@@ -1,5 +1,5 @@
-#ifndef SRC_PLANNING_INCLUDE_UTILS_SMOOTHING_CONFIG_HPP_
-#define SRC_PLANNING_INCLUDE_UTILS_SMOOTHING_CONFIG_HPP_
+#ifndef SRC_PLANNING_INCLUDE_CONFIG_SMOOTHING_CONFIG_HPP_
+#define SRC_PLANNING_INCLUDE_CONFIG_SMOOTHING_CONFIG_HPP_
 
 /**
  * @brief struct for the configuration of the smoothing algorithm
@@ -26,4 +26,4 @@ struct SmoothingConfig {
       : precision(precision), order(order), coeffs_ratio(coeffs_ratio) {}
 };
 
-#endif  // SRC_PLANNING_INCLUDE_UTILS_SMOOTHING_CONFIG_HPP_
+#endif  // SRC_PLANNING_INCLUDE_CONFIG_SMOOTHING_CONFIG_HPP_
