+++ src/planning/include/planning/planning.hpp
@@ -17,13 +17,26 @@
 #include "custom_interfaces/msg/point2d.hpp"
 #include "custom_interfaces/msg/point_array.hpp"
 #include "custom_interfaces/msg/vehicle_state.hpp"
+
+// #include "planning/cone_coloring.hpp"
+// #include "planning/local_path_planner.hpp"
+// #include "planning/path_smoothing.hpp"
+
 #include "planning/cone_coloring.hpp"
-#include "planning/local_path_planner.hpp"
-#include "planning/path_smoothing.hpp"
+#include "planning/outliers.hpp"
+#include "planning/path_calculation.hpp"
+#include "planning/smoothing.hpp"
+
 #include "sensor_msgs/msg/point_cloud2.hpp"
 #include "utils/files.hpp"
 #include "utils/message_converter.hpp"
-#include "utils/pose.hpp"
+// #include "utils/pose.hpp"
+
+#include "common_lib/structures/path_point.hpp"
+#include "common_lib/structures/pose.hpp"
+
+using PathPoint = common_lib::structures::PathPoint;
+using Pose = common_lib::structures::Pose;
 
 using std::placeholders::_1;
 
@@ -40,7 +53,12 @@ class Adapter;
 class Planning : public rclcpp::Node {
   common_lib::competition_logic::Mission mission =
       common_lib::competition_logic::Mission::NONE;              /**< Current planning mission */
-  LocalPathPlanner *local_path_planner = new LocalPathPlanner(); /**< Local path planner instance */
+  //LocalPathPlanner *local_path_planner = new LocalPathPlanner(); /**< Local path planner instance */
+  ConeColoring cone_coloring;
+  Outliers outliers;
+  PathCalculation path_calculation;
+  PathSmoothing path_smoothing;
+
   Adapter *_adapter_;
   std::string mode;
 
@@ -64,7 +82,7 @@ class Planning : public rclcpp::Node {
   bool using_simulated_se_ = false;
   bool recieved_first_track_ = false;
   bool recieved_first_pose_ = false;
-  std::vector<Cone *> cone_array_;
+  std::vector<Cone> cone_array_;
   /**< Subscription to vehicle localization */
   rclcpp::Subscription<custom_interfaces::msg::VehicleState>::SharedPtr vl_sub_;
   /**< Subscription to track map */
@@ -115,7 +133,7 @@ class Planning : public rclcpp::Node {
    * created PointArray message.
    */
 
-  void publish_track_points(const std::vector<PathPoint *> &path) const;
+  void publish_track_points(const std::vector<PathPoint> &path) const;
   /**
    * @brief Publishes predictive track points.
    * @details Depending on the selected mission, this function publishes
@@ -132,12 +150,12 @@ class Planning : public rclcpp::Node {
    * @param after_triangulations_path path after triangulations
    * @param final_path final path after smoothing
    */
-  void publish_visualization_msgs(const std::vector<Cone *> &left_cones,
-                                  const std::vector<Cone *> &right_cones,
-                                  const std::vector<PathPoint *> &after_triangulations_path,
-                                  const std::vector<PathPoint *> &final_path,
-                                  const std::vector<Cone *> &after_rem_blue_cones,
-                                  const std::vector<Cone *> &after_rem_yellow_cones);
+  void publish_visualization_msgs(const std::vector<Cone> &left_cones,
+                                  const std::vector<Cone> &right_cones,
+                                  const std::vector<Cone> &after_refining_blue_cones,
+                                  const std::vector<Cone> &after_refining_yellow_cones,
+                                  const std::vector<PathPoint> &after_triangulations_path,
+                                  const std::vector<PathPoint> &final_path);
 
   /**
    * @brief Checks if the current mission is predictive.
