--- src/planning/include/utils/cone2.hpp
@@ -1,6 +1,0 @@
-#include <vector>
-#include "common_lib/structures/cone.hpp"
-
-using Cone = common_lib::structures::Cone;
-
-int find_cone(std::vector<Cone> &cones, double x, double y);
